+++ src/plugins/core/metaData/keepStreamingMetaDataHandler.ts
@@ -9,25 +9,10 @@ export function keepStreamingMetaDataHandler(type: string, _value: string | unde
     const id = `keep_streaming_${uuid()}`;
     context.httpRegion.hooks.onStreaming.addHook('keepStreaming', async (context: models.ProcessorContext) => {
       if (context.request) {
-        const streamSession: models.UserSession = {
-          id,
-          type: 'Stream',
-          title: `${context.request.method} ${context.request.url}`,
-          description: 'Pending Stream',
-          details: context.request.headers || {},
-        };
         utils.report(context, 'stream until manual cancellation');
-
         await new Promise(resolve => {
-          userSessionStore.setUserSession(streamSession);
-          const dispose = context.progress?.register?.(() => userSessionStore.removeUserSession(id));
-
-          streamSession.delete = () => {
-            dispose?.();
-            resolve(true);
-          };
           if (context.requestClient) {
-            context.requestClient.on('disconnect', () => streamSession.delete?.());
+            context.requestClient.on('disconnect', () => resolve(true));
           }
         });
       }